Move onlyif_cmds_exist decorator to test-specific utils
