+++ nbconvert/utils/tests/test_pandoc.py
@@ -12,7 +12,7 @@
 import os
 import warnings
 
-from ..io import onlyif_cmds_exist
+from ...tests.utils import onlyif_cmds_exist
 
 from nbconvert.tests.base import TestsBase
 from .. import pandoc
@@ -20,6 +20,7 @@ from .. import pandoc
 #-----------------------------------------------------------------------------
 # Classes and functions
 #-----------------------------------------------------------------------------
+
 class TestPandoc(TestsBase):
     """Collection of Pandoc tests"""
 
@@ -47,11 +48,11 @@ class TestPandoc(TestsBase):
             pandoc.check_pandoc_version()
             pandoc.pandoc("", "markdown", "html")
         self.assertEqual(w, [])
-        
+
     @onlyif_cmds_exist('pandoc')
     def test_minimal_version(self):
         original_minversion = pandoc._minimal_version
-        
+
         pandoc._minimal_version = "120.0"
         with warnings.catch_warnings(record=True) as w:
             assert not pandoc.check_pandoc_version()
@@ -59,7 +60,6 @@ class TestPandoc(TestsBase):
 
         pandoc._minimal_version = pandoc.get_pandoc_version()
         assert pandoc.check_pandoc_version()
-
 
 def pandoc_function_raised_missing(f, *args, **kwargs):
     try:
